feat: allow completing tasks

+++ app/routes/__items/tasks/$taskId.tsx
@@ -2,8 +2,9 @@ import type { ActionArgs, LoaderArgs } from '@remix-run/node';
 import { json, redirect } from '@remix-run/node';
 import { Form, useCatch, useLoaderData } from '@remix-run/react';
 import invariant from 'tiny-invariant';
+import * as paths from '~/paths';
 
-import { deleteTask, getTask } from '~/models/task.server';
+import { deleteTask, getTask, toggleTaskComplete } from '~/models/task.server';
 import { requireUserId } from '~/session.server';
 
 export async function loader({ request, params }: LoaderArgs) {
@@ -21,9 +22,17 @@ export async function action({ request, params }: ActionArgs) {
 	const userId = await requireUserId(request);
 	invariant(params.taskId, 'taskId not found');
 
-	await deleteTask({ userId, id: params.taskId });
+	const data = await request.formData();
+	const intent = data.get('intent');
+	invariant(typeof intent === 'string', 'must provide an intent');
 
-	return redirect('/tasks/inbox');
+	if (intent === 'deleteTask') {
+		await deleteTask({ userId, id: params.taskId });
+	} else if (intent === 'completeTask') {
+		await toggleTaskComplete({ userId, id: params.taskId, done: true });
+	}
+
+	return redirect(paths.inbox({}));
 }
 
 export default function TaskDetailsPage() {
@@ -34,11 +43,23 @@ export default function TaskDetailsPage() {
 			<div className="flex items-center">
 				<h3 className="text-2xl font-bold">{data.task.title}</h3>
 
+				<Form method="post" className="ml-8">
+					<button
+						type="submit"
+						className="rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
+						name="intent"
+						value="completeTask">
+						Complete
+					</button>
+				</Form>
+
 				{data.task.deleted == null && (
-					<Form method="post" className="ml-8">
+					<Form method="post" className="ml-2">
 						<button
 							type="submit"
-							className="rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400">
+							className="rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
+							name="intent"
+							value="deleteTask">
 							Delete
 						</button>
 					</Form>
